+++ DemoApp/Screens/Create Chat/NameGroupViewController.swift
@@ -3,7 +3,6 @@
 //
 
 import Foundation
-import Nuke
 import StreamChat
 import StreamChatCommonUI
 import StreamChatUI
@@ -154,9 +153,7 @@ extension NameGroupViewController: UITableViewDataSource {
         guard let cell = tableView.dequeueReusableCell(withIdentifier: UserCell.reuseIdentifier) as? UserCell else { return UITableViewCell() }
         let user = selectedUsers[indexPath.row]
 
-        if let imageURL = user.imageURL {
-            Nuke.loadImage(with: imageURL, into: cell.avatarView)
-        }
+        Components.default.mediaLoader.loadImage(into: cell.avatarView, from: user.imageURL)
 
         cell.nameLabel.text = user.name
         cell.premiumImageView.isHidden = true